Update style.css
So here's what I changed:
in #external links -- I took out:
vertical-align: left;  (vertical by default is a top / bottom kind of thing. Since it's the first element, the default is on the left.)
float: left (removed because this is what the display handles)
and changed:
display: from inline to inline-block  (we were so close!!)

Main -- removed: float: right; 
changed display: inline to inline-block

Added to Media block:
display: block;    (so that it would force the main down) -- cool?